Fixes styling for create event dialog. Adds form field validation for event name, location, description.

+++ src/app/events/events.component.ts
@@ -6,12 +6,11 @@ import { EventCreateDialogComponent } from './event-create/event-create.componen
 import Event from '../shared/models/event.model';
 
 @Component({
-    selector: 'app-events',
-    templateUrl: './events.component.html',
-    styleUrls: ['./events.component.sass']
+  selector: 'app-events',
+  templateUrl: './events.component.html',
+  styleUrls: ['./events.component.sass'],
 })
 export class EventsComponent implements OnInit {
-
   events: Array<Event> = [];
   today = new Date();
   isAdmin = false;
@@ -21,12 +20,12 @@ export class EventsComponent implements OnInit {
     private activatedRoute: ActivatedRoute,
     public dialog: MatDialog,
     private router: Router
-  ) { }
+  ) {}
 
   ngOnInit(): void {
     this.events = this.eventService.getEvents();
     const username: string = this.activatedRoute.snapshot.params['username'];
-    if (username && username.includes("admin")){
+    if (username && username.includes('admin')) {
       this.isAdmin = true;
     }
   }
@@ -34,24 +33,28 @@ export class EventsComponent implements OnInit {
   openDialog(): void {
     const dialogRef = this.dialog.open(EventCreateDialogComponent, {
       width: '800px',
-      maxHeight: '90vh'
+      maxHeight: '90vh',
     });
 
-    dialogRef.afterClosed().subscribe(result => {
-      console.log('The dialog was closed', result);
+    dialogRef.afterClosed().subscribe((newEvent: Event) => {
+      if (newEvent) {
+        this.events = this.eventService.getEvents();
+      }
     });
   }
 
-  setDateFromToday(offset: number): string{
+  setDateFromToday(offset: number): string {
     let day = this.today.getDate() + offset;
     let month = this.today.getMonth() + 1;
     let year = this.today.getFullYear();
 
-    return day + "/" + month + "/" + year;
+    return day + '/' + month + '/' + year;
   }
 
-  delete(id: number): void{
-    let index = this.events.findIndex(event => event.id === id);
-    this.eventService.deleteEvent(index);
+  delete(id: number | undefined) {
+    if (id) {
+      let index = this.events.findIndex((event) => event.id === id);
+      this.eventService.deleteEvent(index);
+    }
   }
 }
